refactor(knowledge-base): extract KIND_STYLES from SourceRow + freshen empty-state copy
- Move KIND_STYLES + KindStyle interface + their 6 lucide icons into
  SourceRow.kinds.ts to bring SourceRow.tsx (615 → 567 lines) back under
  the quality-gate baseline. The kind/icon mapping is reference data,
  not row behavior, so the sibling module is the more natural home.
- Update KnowledgeBasePage masthead and empty-state copy to cover the
  new "skill manifest" source kind alongside the existing
  Doc/Web/PDF flow — previous wording said "every source is chunked
  and embedded hourly," which is no longer true for manifests
  (addresses CodeRabbit review on PR #1482).
