+++ components/Pages/Admin/KnowledgeBasePage/KnowledgeBasePage.tsx
@@ -178,7 +178,8 @@ function Masthead({
           <p className="mt-1.5 max-w-[68ch] text-sm leading-relaxed text-stone-600 dark:text-zinc-400">
             Documents the{" "}
             <span className="font-medium text-stone-900 dark:text-zinc-100">{communityName}</span>{" "}
-            assistant reads from. Each source is fetched, chunked, and embedded hourly.
+            assistant reads from. Documents are fetched and embedded hourly; skill manifests
+            are fetched and passed to the assistant as instructions.
           </p>
         </div>
 
@@ -337,8 +338,8 @@ function EmptyState({
         No sources yet
       </h2>
       <p className="mx-auto mt-1.5 max-w-[44ch] text-sm leading-relaxed text-stone-600 dark:text-zinc-400">
-        Add a publicly-shared Google Doc, a web page, or a PDF URL and your chatbot will start
-        answering questions from it within the hour.
+        Add a publicly-shared Google Doc, a web page, a PDF URL, or a skill manifest, and your
+        chatbot will start answering questions from it within the hour.
       </p>
 
       <Button
